requirements.txt
keras==2.12.0
numpy==1.24.3
opencv_python_headless==4.7.0.72
Pillow==9.5.0
pytesseract==0.3.10
SpeechRecognition==3.10.0
textblob==0.17.1
